release 0.1.92-patch source code

+++ services/ecs/v2/model/model_server_limits.go
@@ -58,7 +58,7 @@ type ServerLimits struct {
 	TotalInstancesUsed int32 `json:"totalInstancesUsed"`
 
 	// 当前内存使用容量（单位：MB）。
-	TotalRAMUsed int32 `json:"totalRAMUsed"`
+	TotalRAMUsed int64 `json:"totalRAMUsed"`
 
 	// 当前安全组使用个数。
 	TotalSecurityGroupsUsed int32 `json:"totalSecurityGroupsUsed"`
